backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
google-generativeai==0.7.2
scikit-learn==1.5.0
python-dotenv==1.0.1
